+++ Invasion1D/Models/Item.cs
@@ -1,6 +1,6 @@
 ﻿namespace Invasion1D.Models
 {
-	public abstract class Item(Dimension dimension, double position, Color color) : Interactive(dimension, position, color)
+	public abstract class Item(Dimension dimension, float position, Color color) : Interactive(dimension, position, color)
 	{
 		public abstract bool Power(Character character);
 	}
